PR15.6: adopt cow-solana-rpc in the indexer
