+++ crates/solana-indexer/Cargo.toml
@@ -21,10 +21,10 @@ bigdecimal = { workspace = true }
 bytes = { workspace = true }
 clap = { workspace = true }
 configs = { workspace = true }
+cow-solana-rpc = { workspace = true }
 derive_more = { workspace = true }
 futures = { workspace = true }
 humantime-serde = { workspace = true }
-itertools = { workspace = true }
 observe = { workspace = true }
 rustls = { workspace = true }
 serde = { workspace = true }
@@ -34,9 +34,6 @@ settlement-interface = {
   git = "https://github.com/cowprotocol/solana-programs",
   rev = "4d4fc46f594e0159246afaf148f0af2a463d0095"
 }
-solana-client = { workspace = true }
-solana-commitment-config = { workspace = true }
-solana-rpc-client-api = { workspace = true }
 solana-sdk = { workspace = true }
 spl-token-2022-interface = { workspace = true }
 spl-token-interface = { workspace = true }
@@ -51,6 +48,7 @@ yellowstone-grpc-client = { workspace = true }
 yellowstone-grpc-proto = { workspace = true }
 
 [dev-dependencies]
+cow-solana-rpc = { workspace = true, features = ["test-util"] }
 serde_json = { workspace = true }
 settlement-client = {
   package = "settlement-client",
