+++ crates/solana-indexer/src/run.rs
@@ -9,10 +9,10 @@ use {
             ingester::{Error, INGEST_TO_DECODER_CAPACITY, Ingester, Resume},
         },
         persistence::Postgres,
-        rpc::Rpc,
         yellowstone,
     },
     clap::Parser,
+    cow_solana_rpc::{CommitmentConfig, SolanaRPC},
     sqlx::PgPool,
     std::{
         path::PathBuf,
@@ -57,7 +57,12 @@ async fn run(config: Config, start_slot: Option<u64>) {
         .await
         .expect("database connection");
     let persistence = Postgres::new(pool);
-    let rpc = Rpc::new(config.rpc.endpoint, config.rpc.request_timeout);
+    // Confirmed commitment, matching the stream subscription.
+    let rpc = SolanaRPC::new_with_timeout_and_commitment(
+        &config.rpc.endpoint,
+        config.rpc.request_timeout,
+        CommitmentConfig::confirmed(),
+    );
 
     let (tx, rx) = mpsc::channel(INGEST_TO_DECODER_CAPACITY);
     let settlement_program = config.chain.settlement_program_id;
