+++ crates/solana-indexer/src/types/errors.rs
@@ -38,7 +38,7 @@ pub(crate) enum PersistenceError {
     Database(#[from] sqlx::Error),
     /// The account lookup enriching a batch failed.
     #[error("rpc: {0}")]
-    Rpc(#[from] solana_client::client_error::ClientError),
+    Rpc(#[from] cow_solana_rpc::Error),
 }
 
 /// Failures surfaced from the stream boundary.
